+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.8.1 Feb 11, 2021
+
+### New
+
+- `tuple_list_as_array` parameter in `tvm.run_get` function which controls lists representation. 
+Default is stack-like based on nested tuples. If set to `true` then returned lists are encoded as plain arrays.
+This reduces stack size requirements for long lists.
+    
 ## 1.8.0 Feb 11, 2021
 
 ### New